backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dj-database-url==2.0.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.6
gunicorn==21.2.0